Modifications in the muon block
